Fixed Windows file and folder renaming issue.
In previous commits, I thought renaming the filename because
of casing conflicts between Windows and Linux will solve the issue.
Windows git doesn't detects the changes if I just simply renamed
the folder or filename if the only changes is the casing.
I renamed the file entirely to detect it.
