+++ src/components/index.js
@@ -2,4 +2,4 @@ export { default as MovieInformation } from './MovieInformation/MovieInformation
 export { default as Movie } from './Movie/Movie';
 export { default as Actors } from './Actors/Actors';
 export { default as Profile } from './Profile/Profile';
-export { default as NavBar } from './NavBar/NavBar';
+export { default as NavBar } from './NavBarTemp/NavBarTemp';
